python/requirements.txt
mysql-connector-python
sentence-transformers
numpy
faiss-cpu
ollama
fastapi
pydantic
uvicorn
